- always store the memory block size allocated with g_slice_alloc[0] to a   variable and use that for freeing - fixed wrong memory block size in find_grain_positions()
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code@13761 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ branches/GWYDDION-3/libgwy/mask-field-grains.c
@@ -271,7 +271,7 @@ find_grain_positions(const GwyMaskField *field)
     MaskField *priv = field->priv;
     guint ngrains = priv->ngrains;
     guint xres = field->xres, yres = field->yres;
-    guint memsize = (ngrains + 1)*sizeof(guint);
+    guint memsize = (ngrains + 1)*sizeof(gdouble);
 
     gdouble *xc = g_slice_alloc0(memsize),
             *yc = g_slice_alloc0(memsize),
